Extract schedule logic into reusable composable with caching and tests
