+++ frontend/src/views/ScheduleView.vue
@@ -21,209 +21,14 @@
 </template>
 
 <script setup>
-import {
-  computed,
-  onMounted,
-  onBeforeUnmount,
-  shallowRef,
-  watch,
-  defineAsyncComponent
-} from 'vue';
-import { storeToRefs } from 'pinia';
-import { useScheduleStore } from '../store/schedule';
+import { defineAsyncComponent } from 'vue';
+import { useSchedule } from '../composables/useSchedule';
 
 const GameRow = defineAsyncComponent(() =>
   import('../components/GameRow.vue')
 );
 
-const scheduleStore = useScheduleStore();
-const { schedule } = storeToRefs(scheduleStore);
-
-const scheduleCache = new Map();
-
-let controller;
-
-let refreshInterval;
-
-const today = new Date().toISOString().slice(0, 10);
-
-// Cache of Intl.DateTimeFormat instances by locale
-const formatterCache = new Map();
-// Cache formatted date results keyed by input date string + locale
-const formattedDateCache = new Map();
-const TIME_ZONE = 'America/New_York';
-const nyHourFormatter = new Intl.DateTimeFormat('en-US', {
-  timeZone: TIME_ZONE,
-  hour: '2-digit',
-  hour12: false
-});
-
-const games = shallowRef([]);
-
-watch(
-  schedule,
-  (scheduleVal) => {
-    games.value = Array.isArray(scheduleVal)
-      ? scheduleVal.flatMap((d) => d?.games || [])
-      : [];
-  },
-  { immediate: true }
-);
-
-const currentDate = computed(() => {
-  const scheduleVal = schedule.value;
-  const dateStr =
-    scheduleVal[0]?.date ||
-    // Use optional chaining for array access to avoid errors when the
-    // schedule array is empty.
-    scheduleVal[0]?.games?.[0]?.gameDate;
-  return dateStr ? dateStr.slice(0, 10) : today;
-});
-
-const headerDate = computed(() =>
-  currentDate.value === today
-    ? `Today - ${formatDate(currentDate.value)}`
-    : formatDate(currentDate.value)
-);
-
-async function fetchSchedule(
-  dateStr,
-  { cacheOnly = false, prefetch = true, force = false } = {},
-) {
-  if (!cacheOnly) {
-    if (controller) controller.abort();
-    controller = new AbortController();
-  }
-
-  if (!force && scheduleCache.has(dateStr)) {
-    const cached = scheduleCache.get(dateStr);
-    if (!cacheOnly) scheduleStore.setSchedule(cached);
-    if (prefetch && !cacheOnly) prefetchAdjacent(dateStr);
-    return cached;
-  }
-
-  if (prefetch && !cacheOnly) prefetchAdjacent(dateStr);
-
-  const options = {};
-  if (!cacheOnly) options.signal = controller.signal;
-  const resp = await fetch(`/api/schedule/?date=${dateStr}`, options);
-  const data = await resp.json();
-  scheduleCache.set(dateStr, data);
-  if (!cacheOnly) scheduleStore.setSchedule(data);
-
-  // Fetch win probability predictions for each game in parallel
-  // const predictionPromises = [];
-  // for (const day of schedule.value) {
-  //   for (const game of day.games || []) {
-  //     const p = fetch(`/api/games/${game.gamePk}/prediction/`)
-  //       .then((r) => (r.ok ? r.json() : null))
-  //       .then((data) => {
-  //         if (data) game.prediction = data;
-  //       })
-  //       .catch(() => {});
-  //     predictionPromises.push(p);
-  //   }
-  // }
-  // await Promise.all(predictionPromises);
-
-  return data;
-}
-
-function prefetchAdjacent(dateStr) {
-  const date = new Date(dateStr);
-  const prev = new Date(date);
-  const next = new Date(date);
-  prev.setDate(prev.getDate() - 1);
-  next.setDate(next.getDate() + 1);
-  const prevIso = prev.toISOString().split('T')[0];
-  const nextIso = next.toISOString().split('T')[0];
-  fetchSchedule(prevIso, { cacheOnly: true, prefetch: false });
-  fetchSchedule(nextIso, { cacheOnly: true, prefetch: false });
-}
-
-async function prevDay() {
-  const date = new Date(currentDate.value);
-  date.setDate(date.getDate() - 1);
-  const iso = date.toISOString().split('T')[0];
-  try {
-    await fetchSchedule(iso, { force: iso === today });
-  } catch {}
-}
-
-async function nextDay() {
-  const date = new Date(currentDate.value);
-  date.setDate(date.getDate() + 1);
-  const iso = date.toISOString().split('T')[0];
-  try {
-    await fetchSchedule(iso, { force: iso === today });
-  } catch {}
-}
-
-function getFormatter(locale) {
-  const loc = locale || undefined;
-  if (!formatterCache.has(loc)) {
-    formatterCache.set(
-      loc,
-      new Intl.DateTimeFormat(loc, {
-        weekday: 'long',
-        month: 'short',
-        day: 'numeric',
-        timeZone: TIME_ZONE
-      })
-    );
-  }
-  return formatterCache.get(loc);
-}
-
-function formatDate(dateStr, locale) {
-  if (!dateStr) return '';
-  const cacheKey = `${dateStr}|${locale || ''}`;
-  if (formattedDateCache.has(cacheKey)) return formattedDateCache.get(cacheKey);
-  let date;
-  if (/^\d{4}-\d{2}-\d{2}$/.test(dateStr)) {
-    try {
-      const [y, m, d] = dateStr.split('-').map(Number);
-      const noonUTC = Date.UTC(y, m - 1, d, 12, 0, 0);
-      const hourInNY = Number(nyHourFormatter.format(noonUTC));
-      const offset = 12 - hourInNY;
-      const offsetStr = String(Math.abs(offset)).padStart(2, '0');
-      const sign = offset >= 0 ? '-' : '+';
-      date = new Date(`${dateStr}T00:00:00${sign}${offsetStr}:00`);
-    } catch {
-      date = new Date(`${dateStr}T00:00:00-05:00`);
-    }
-  } else {
-    date = new Date(dateStr);
-  }
-  const formatted = getFormatter(locale).format(date);
-  formattedDateCache.set(cacheKey, formatted);
-  return formatted;
-}
-
-watch(
-  currentDate,
-  (newDate) => {
-    if (refreshInterval) {
-      clearInterval(refreshInterval);
-      refreshInterval = null;
-    }
-    if (newDate === today) {
-      refreshInterval = setInterval(() => {
-        fetchSchedule(today, { force: true });
-      }, 60000);
-    }
-  },
-  { immediate: true }
-);
-
-onMounted(() => {
-  fetchSchedule(today, { force: true });
-});
-
-onBeforeUnmount(() => {
-  if (refreshInterval) clearInterval(refreshInterval);
-});
-
+const { schedule, games, headerDate, prevDay, nextDay } = useSchedule();
 </script>
 
 <style scoped>
@@ -282,3 +87,4 @@ onBeforeUnmount(() => {
   gap: 1rem;
 }
 </style>
+
